make admission for student
